some norminette compliance and organization
