+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/05 16:01:34 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/10/12 16:52:07 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/10/12 17:13:06 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -102,14 +102,35 @@ void					unset_env(t_list **env, char *str);
 void					unset_exp(t_list **exp, char *str);
 void					unset(t_list **env, t_list **exp, char **arr);
 
+//++++++++++++++++ error_handling/[.....] +++++++++++++++++++++++++++++++++++++
+// errors.c
+int						syntax_error(t_minishell *ms);
+int						token_message(char token);
+void					malloc_error(t_minishell *ms);
+
+// syntax_errors.c
+int						begin_error(char *str);
+int						end_of_string_error(char *str);
+int						redir_error(char *str);
+int						double_redir_error(char *str, char c);
+int						sucession_error(char *str);
+
+// syntax_errors2.c
+int						quote_error(char *str);
+int						pipe_error(char *str);
+int						dollar_error(char *str);
+
+
 //++++++++++++++++ exec/[.....] +++++++++++++++++++++++++++++++++++++++++++++++
+// exec_built_ins.c
+int						is_built_in(char *str);
+void					built_ins(t_minishell *ms, char **cmd_with_flags, int exit);
+void					exp_env_unset(t_minishell *ms, char **cmd_with_flags);
+
 // exec_utils.c
 char					**path_init(t_minishell *ms);
 char					*is_exec(char *str, char **paths);
-int						is_built_in(char *str);
-void					built_ins(t_minishell *ms, char **cmd_with_flags, int exit);
 void					last_cmd(t_minishell *ms, t_cmdlist *cmdlist, int i);
-void					exp_env_unset(t_minishell *ms, char **cmd_with_flags);
 
 // exec.c
 void					exec(t_minishell *ms, t_cmdlist *cmdlist);
@@ -174,11 +195,11 @@ void					arr_print(char *str, char **arr);
 // env_utl.c
 char					*env_var_str(char *str, t_list **env);
 void					env_var(t_list **env, char **arr);
+char					*path_str(t_list *env);
 
 // frees.c
 void					free_ms(t_minishell *ms);
 int						free_array(char **arr);
-void					malloc_error(t_minishell *ms);
 void					free_cmd_list(t_cmdlist *cmdlist);
 void					free_list_malloc(t_list **exp);
 
@@ -194,20 +215,6 @@ int						strlen_chr(char *str, char c);
 int						strcmp_chr(char *s1, char *s2, char c);
 
 // +++++++++++++++ ./[.....] ++++++++++++++++++++++++++++++++++++++++++++++++++
-// errors.c
-int						quote_error(char *str);
-int						pipe_error(char *str);
-int						dollar_error(char *str);
-int						token_message(char token);
-int						syntax_error(t_minishell *ms);
-
-// errors2.c
-int						begin_error(char *str);
-int						end_of_string_error(char *str);
-int						redir_error(char *str);
-int						double_redir_error(char *str, char c);
-int						sucession_error(char *str);
-
 // prompt.c
 char					*set_prompt(t_minishell *ms);
