+++ src/error_handling/syntax_errors2.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   errors.c                                           :+:      :+:    :+:   */
+/*   syntax_errors2.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/04 16:57:08 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/05 16:24:00 by mortins-         ###   ########.fr       */
+/*   Created: 2023/10/12 17:05:56 by mortins-          #+#    #+#             */
+/*   Updated: 2023/10/12 17:06:29 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
 // Checks for unclosed quote marks
 int	quote_error(char *str)
@@ -80,32 +80,6 @@ int	dollar_error(char *str)
 			return (token_message('$'));
 		else
 			i++;
-	}
-	return (0);
-}
-
-int	token_message(char token)
-{
-	int	fd;
-	fd = dup(STDOUT_FILENO);
-	dup2(STDERR_FILENO, STDOUT_FILENO);
-	printf("MiniShell: syntax error near unexpected token '%c'\n", token);
-	dup2(fd, STDOUT_FILENO);
-	close(fd);
-	return (1);
-}
-
-int	syntax_error(t_minishell *ms)
-{
-	if (quote_error(ms->str) || pipe_error(ms->str) || begin_error(ms->str) || \
-		end_of_string_error(ms->str) || double_redir_error(ms->str, '<') || \
-		double_redir_error(ms->str, '>') || dollar_error(ms->str) || \
-		sucession_error(ms->str) || redir_error(ms->str))
-	{
-		free(ms->prompt);
-		free(ms->str);
-		g_exit = 2;
-		return (1);
 	}
 	return (0);
 }
