+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/05 15:59:22 by ddiniz-m          #+#    #+#             */
-/*   Updated: 2023/10/12 15:46:46 by mortins-         ###   ########.fr       */
+/*   Updated: 2023/10/12 16:54:22 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,8 @@ int	main(int ac, char **av, char **envp)
 // `>output>>append echo 1 2 3 <input<<heredoc | <in<<here ls -l >out>>app`
 
 /*	To do:
-	- `export VAR` should create an environment variable VAR when there isn't already one
+	- `export VAR` should create an environment variable VAR when there isn't
+		already one
 	- Return error message when command is invalid
 	- Figure out where/how we are gonna substitute ($VAR) by it's actual value
  */
