+++ src/utils/frees.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   frees.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ddiniz-m <ddiniz-m@student.42.fr>          +#+  +:+       +#+        */
+/*   By: mortins- <mortins-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/23 16:53:12 by mortins-          #+#    #+#             */
-/*   Updated: 2023/10/06 12:49:41 by ddiniz-m         ###   ########.fr       */
+/*   Updated: 2023/10/12 17:09:48 by mortins-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,13 +38,6 @@ int	free_array(char **arr)
 	if (arr)
 		free(arr);
 	return (0);
-}
-
-void	malloc_error(t_minishell *ms)
-{
-	write(2, "Error: Malloc failed\n", 22);
-	g_exit = 12;
-	free_ms(ms);
 }
 
 void	free_list(t_list **list)
